requirements.txt
flask==2.3.3
flask-cors==4.0.0
pybind11==2.11.1
fasttext-wheel==0.9.2
gunicorn==21.2.0
requests==2.31.0
